Phase 5: Remove original dives.py file and optimize modular structure
- Successfully removed original 130KB dives.py file (3,421 lines)
- All 42 functions moved to focused modular structure
- Updated test imports to use new module locations
- Found and restored missing parse_dive_element function from git history
- Added parse_dive_element function to dives_import.py module
- Fixed all test imports to use correct module locations
- Consolidated 4 duplicate functions into dives_utils.py (DRY principle)
- Restored original advanced calculate_similarity with multi-algorithm approach
- Restored original advanced find_dive_site_by_import_id with similarity matching
- Eliminated code duplication while maintaining full functionality
- Verified no remaining references to old file
- Confirmed all 24 routes working correctly
- Main app imports successfully (134 total routes)
- Advanced dive site matching capabilities fully preserved

Phase: REMOVE_ORIGINAL | Status: Complete
